move to private socket header
