+++ source/posix/socket.c
@@ -14,6 +14,7 @@
 #include <aws/io/event_loop.h>
 #include <aws/io/logging.h>
 #include <aws/io/private/event_loop_impl.h>
+#include <aws/io/private/socket_impl.h>
 
 #include <arpa/inet.h>
 #include <aws/io/io.h>
